Extract the content caching decorators to content/utils/cache.py
Both this issue and #14300 need them importable from a content-side
module once the public viewsets move out of kolibri.core.public.
