+++ kolibri/core/content/api.py
@@ -1,4 +1,3 @@
-import hashlib
 import logging
 import re
 from base64 import urlsafe_b64decode
@@ -17,15 +16,11 @@ from django.db.models.aggregates import Count
 from django.http import Http404
 from django.http import HttpResponse
 from django.urls import reverse
-from django.utils.cache import add_never_cache_headers
 from django.utils.decorators import method_decorator
-from django.utils.encoding import force_bytes
-from django.utils.encoding import iri_to_uri
 from django.utils.text import smart_split
 from django.utils.text import unescape_string_literal
 from django.views import View
 from django.views.decorators.cache import cache_page
-from django.views.decorators.cache import never_cache
 from django.views.decorators.http import etag
 from django_filters.rest_framework import BaseInFilter
 from django_filters.rest_framework import BooleanFilter
@@ -62,7 +57,6 @@ from kolibri.core.api import ListModelMixin
 from kolibri.core.api import ReadOnlyValuesViewset
 from kolibri.core.api import ValuesMethodField
 from kolibri.core.api import ValuesViewsetOrderingFilter
-from kolibri.core.auth.middleware import session_exempt
 from kolibri.core.auth.permissions import KolibriAuthPermissions
 from kolibri.core.auth.permissions import KolibriAuthPermissionsFilter
 from kolibri.core.bookmarks.models import Bookmark
@@ -76,6 +70,12 @@ from kolibri.core.content.models import ContentRequestReason
 from kolibri.core.content.models import ContentRequestStatus
 from kolibri.core.content.permissions import CanManageContent
 from kolibri.core.content.tasks import automatic_user_imported_resource_cleanup
+from kolibri.core.content.utils.cache import get_cache_key
+from kolibri.core.content.utils.cache import get_course_ids
+from kolibri.core.content.utils.cache import no_cache_on_method
+from kolibri.core.content.utils.cache import REMOTE_ETAG_CACHE_KEY
+from kolibri.core.content.utils.cache import remote_metadata_cache
+from kolibri.core.content.utils.cache import REMOTE_URL_PARAM
 from kolibri.core.content.utils.content_types_tools import (
     renderable_contentnodes_q_filter,
 )
@@ -95,7 +95,6 @@ from kolibri.core.content.utils.paths import get_local_content_storage_file_url
 from kolibri.core.content.utils.paths import get_v2_channel_lookup_url
 from kolibri.core.content.utils.search import get_available_metadata_labels
 from kolibri.core.content.utils.stopwords import stopwords_set
-from kolibri.core.device.models import ContentCacheKey
 from kolibri.core.device.permissions import FromAppContextPermission
 from kolibri.core.discovery.utils.network.client import NetworkClient
 from kolibri.core.discovery.utils.network.errors import NetworkClientError
@@ -119,93 +118,6 @@ from kolibri.utils.conf import OPTIONS
 from kolibri.utils.version import version_matches_range
 
 logger = logging.getLogger(__name__)
-
-
-REMOTE_ETAG_CACHE_KEY = "remote_content_etag_{}"
-
-REMOTE_URL_PARAM = "baseurl"
-
-
-def get_cache_key(*args, **kwargs):
-    return str(ContentCacheKey.get_cache_key())
-
-
-def get_course_ids():
-    cache_key = "COURSE_IDS_{}".format(get_cache_key())
-    cached_data = cache.get(cache_key)
-    if cached_data is not None:
-        return cached_data
-    updated_data = list(
-        ContentNode.objects.filter(
-            available=True, modality=modalities.COURSE
-        ).values_list("id", flat=True)
-    )
-    cache.set(cache_key, updated_data, 3600)
-    return updated_data
-
-
-def metadata_cache(view_func, cache_key_func=get_cache_key):
-    """
-    Decorator to apply an Etag sensitive page cache
-    """
-
-    @etag(cache_key_func)
-    def wrapper_func(*args, **kwargs):
-        try:
-            request = args[0]
-            request = kwargs.get("request", request)
-        except IndexError:
-            request = kwargs.get("request", None)
-        # Prevent the Django caching middleware from caching
-        # this response, as we want to cache it ourselves
-        request._cache_update_cache = False
-        key_prefix = cache_key_func(request)
-        url_key = hashlib.md5(
-            force_bytes(iri_to_uri(request.build_absolute_uri()))
-        ).hexdigest()
-        response = None
-        if key_prefix is not None:
-            cache_key = "{}:{}".format(key_prefix, url_key)
-            response = cache.get(cache_key)
-        if response is None:
-            response = view_func(*args, **kwargs)
-            if response.status_code == status.HTTP_200_OK:
-                if key_prefix is None:
-                    key_prefix = cache_key_func(request)
-                if (
-                    key_prefix is not None
-                    and hasattr(response, "render")
-                    and callable(response.render)
-                ):
-                    cache_key = "{}:{}".format(key_prefix, url_key)
-                    response.add_post_render_callback(
-                        lambda r: cache.set(cache_key, r, timeout=3600)
-                    )
-            else:
-                # Don't cache responses that returned an error code
-                add_never_cache_headers(response)
-        return response
-
-    return wrapper_func
-
-
-def get_remote_cache_key(request, *args, **kwargs):
-    if REMOTE_URL_PARAM in request.GET:
-        return cache.get(REMOTE_ETAG_CACHE_KEY.format(request.GET[REMOTE_URL_PARAM]))
-    return get_cache_key(*args, **kwargs)
-
-
-def remote_metadata_cache(view_func):
-    return session_exempt(
-        metadata_cache(view_func, cache_key_func=get_remote_cache_key)
-    )
-
-
-def no_cache_on_method(view_func):
-    """
-    Decorator to disable caching for a particular method
-    """
-    return method_decorator(never_cache, name="dispatch")(view_func)
 
 
 class RemoteMixin:
